+++ src/app/organizations/[name]/offers/page.tsx
@@ -3,6 +3,7 @@ import { PageBox } from '@/app/page-box';
 import { Footer } from '@/components/common/footer';
 import { TopBar } from '@/components/common/top-bar';
 import { OfferContainer } from '@/app/organizations/[name]/offers/offer-container';
+import { OfferForm } from '@/app/organizations/[name]/offers/offer-form';
 
 interface OrganizationOffersPageProps {
     params: Record<string, any>;
@@ -15,6 +16,7 @@ export default function OrganizationOffersPage({ params }: OrganizationOffersPag
         <main className={styles['organization-offers-page']}>
             <TopBar />
             <PageBox>
+                <OfferForm organizationName={name} />
                 <OfferContainer organizationName={name} />
                 <Footer />
             </PageBox>