+++ src/app/settings/[user]/[object]/[action]/add-categories-creator.tsx
@@ -31,7 +31,7 @@ export function AddCategoriesCreator() {
         <form className={styles['creator-form']}>
             <InputString label={'Category name'} name={'category-name'} value={categoryName} onChange={onCategoryNameChange} />
             <div className={styles['navigation-box']}>
-                <SubmitButton label={'Submit'} link={'/'} size="small" onSubmit={onSubmit} />
+                <SubmitButton label={'Submit'} size="small" onSubmit={onSubmit} />
             </div>
         </form>
     );