+++ src/app/api/api.ts
@@ -294,7 +294,7 @@ export const Organization = {
         return res.text();
     },
     /** Retrieves organization image with the given name [Access: Logged in users]
-     * 
+     *
      * organization - Targeted organization
      */
     async getOrganizationImageEnd(organization: string): Promise<AttachmentTypes.FileModel> {
@@ -493,7 +493,7 @@ export const Categories = {
 };
 export const Offer = {
     /** Retrieves every shop from the system [Access: Logged in users]
-     * 
+     *
      * category - Optional targeted category
      */
     async getShopsEnd(category?: string): Promise<OfferTypes.ShopModel[]> {
@@ -513,7 +513,7 @@ export const Offer = {
         return res.json();
     },
     /** Retrieves offers from a given organization [Access: Logged in users]
-     * 
+     *
      * organization - Targeted organization
      */
     async getOffersEnd(organization: string, category?: string): Promise<OfferTypes.ShopOfferModel[]> {
@@ -533,7 +533,7 @@ export const Offer = {
         return res.json();
     },
     /** Adds new offer to the system [Access: Administrator]
-     * 
+     *
      * offer - New offer
      */
     async addOfferEnd(offer: OfferTypes.ShopOfferModel): Promise<number> {
@@ -554,7 +554,7 @@ export const Offer = {
         return parseInt(await res.text());
     },
     /** Changes details of an offer [Access: Administrator]
-     * 
+     *
      * offer - Offer object
      */
     async changeOfferEnd(offer: OfferTypes.ShopOfferModel): Promise<string> {
@@ -575,7 +575,7 @@ export const Offer = {
         return res.text();
     },
     /** Deletes targeted offer [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
      */
     async deleteOfferEnd(ID: number): Promise<string> {
@@ -595,9 +595,9 @@ export const Offer = {
         return res.text();
     },
     /** Sets offer discount details [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
-     * 
+     *
      * discount - Object with discount details (null means clear discount)
      */
     async setOfferDiscountEnd(ID: number, discount?: OfferTypes.ShopDiscountModel): Promise<string> {
@@ -618,7 +618,7 @@ export const Offer = {
         return res.text();
     },
     /** Retrieves offer image with the given ID [Access: Logged in users]
-     * 
+     *
      * ID - Offer ID
      */
     async getOfferImageEnd(ID: number): Promise<AttachmentTypes.FileModel> {
@@ -643,12 +643,12 @@ export const Offer = {
         };
     },
     /** Saves image for a given offer [Access: Administrator]
-     * 
+     *
      * file - Form file with the offer image
-     * 
+     *
      * ID - Targeted offer ID
      */
-    async setOfferImageEnd(ID: string, image: FormData): Promise<string> {
+    async setOfferImageEnd(ID: number, image: FormData): Promise<string> {
         const res = await fetch(`${API_BASE_URL}/Offer/SetOfferImage/${ID}`, {
             mode: 'cors',
             method: 'PUT',
@@ -666,7 +666,7 @@ export const Offer = {
         return res.text();
     },
     /** Deletes image from a given offer [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
      */
     async deleteOfferImageEnd(ID: number): Promise<string> {
@@ -686,7 +686,7 @@ export const Offer = {
         return res.text();
     },
     /** Retrieves list of every code assigned to an offer [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
      */
     async checkOfferCodesEnd(ID: number): Promise<OfferTypes.CodeModel[]> {
@@ -706,9 +706,9 @@ export const Offer = {
         return res.json();
     },
     /** Adds codes to a specified offer [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
-     * 
+     *
      * codes - Array of codes of NewCodeModel schema
      */
     async addCodesEnd(ID: number, codes: OfferTypes.NewCodeModel[]): Promise<string> {
@@ -729,9 +729,9 @@ export const Offer = {
         return res.text();
     },
     /** Adds codes from a file to a specified offer [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
-     * 
+     *
      * fileCodes - File with codes formatted in JSON with NewCodeModel schema
      */
     async addCodesFromFileEnd(ID: number, fileCodes: FormData): Promise<string> {
@@ -752,9 +752,9 @@ export const Offer = {
         return res.text();
     },
     /** Changes current state of a given code [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
-     * 
+     *
      * code - Targeted code
      */
     async changeCodeStateEnd(ID: number, code?: number): Promise<string> {
@@ -775,9 +775,9 @@ export const Offer = {
         return res.text();
     },
     /** Removes a given code that is not already redeemed [Access: Administrator]
-     * 
+     *
      * ID - Targeted offer ID
-     * 
+     *
      * code - Targeted code
      */
     async deleteCodeEnd(ID: number, code?: number): Promise<string> {
@@ -817,7 +817,7 @@ export const Transactions = {
         return res.json();
     },
     /** Retrieves every transaction in the system [Access: Administrator]
-     * 
+     *
      * organization - Optional parameter that specifies which shop will be included in the list
      */
     async getAllTransactionsEnd(organization?: string): Promise<TransactionModel[]> {
@@ -837,7 +837,7 @@ export const Transactions = {
         return res.json();
     },
     /** Retrieves information about code availability in a given offer [Access: Logged in users]
-     * 
+     *
      * offerID - Targeted offer ID
      */
     async isCodeAvailable(offerID: number): Promise<boolean> {
@@ -857,7 +857,7 @@ export const Transactions = {
         return res.json();
     },
     /** Buys a code from a given offer for the current user [Access: Logged in users]
-     * 
+     *
      * offerID - Targeted offer ID
      */
     async buyCode(offerID: number): Promise<OfferTypes.NewCodeModel> {