+++ src/app/organizations/[name]/offers/[id]/page.tsx
@@ -3,6 +3,7 @@ import { TopBar } from '@/components/common/top-bar';
 import { PageBox } from '@/app/page-box';
 import { Footer } from '@/components/common/footer';
 import { OfferCodeContainer } from '@/app/organizations/[name]/offers/[id]/offer-id-container';
+import { OfferCodeForm } from '@/app/organizations/[name]/offers/[id]/offer-code-form';
 
 interface OfferIdPageProps {
     params: Record<string, any>;
@@ -15,7 +16,7 @@ export default function OfferIdPage({ params }: OfferIdPageProps) {
         <main className={styles['offer-id-page']}>
             <TopBar />
             <PageBox>
-                {/*<OfferCodeForm offerId={id} />*/}
+                <OfferCodeForm offerId={id} />
                 <OfferCodeContainer offerId={Number(id)} />
                 <Footer />
             </PageBox>