+++ src/app/browse-products.tsx
@@ -3,6 +3,7 @@
 import styles from '@/styles/app/page.module.scss';
 import { useCategories } from '@/hooks/use-categories';
 import { CategoryCart } from '@/app/category-cart';
+import { Loader } from '@/components/common/loader';
 
 export function BrowseProducts() {
     const { isLoading, categories } = useCategories();
@@ -12,7 +13,7 @@ export function BrowseProducts() {
             <h2 className={styles['browse-products__header']}>Browse products</h2>
             <div>
                 {isLoading
-                    ? 'Loading categories...'
+                    ? <Loader />
                     : (
                         <div className={styles['browse-products__container']}>
                             {categories.map(category => <CategoryCart category={category} key={category.name} />)}