+++ src/components/common/loader.tsx
@@ -1,8 +1,12 @@
 import styles from '@/styles/components/common/loader.module.scss';
 
-export function Loader() {
+interface LoaderProps {
+    isAbsolute?: boolean;
+}
+
+export function Loader({ isAbsolute = false }: LoaderProps) {
     return (
-        <div className={styles['loader-component']}>
+        <div className={styles['loader-component']} data-is-absolute={isAbsolute}>
             <div className={styles['loader-component__background']}></div>
             <span className={styles['loader']}></span>
         </div>