+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Marek Kurańda](https://github.com/mjkuranda): `/manage` subpage.
 - [Marek Kurańda](https://github.com/mjkuranda): Creators for `categories`.
 - [Marek Kurańda](https://github.com/mjkuranda): `InputNumber` component.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputImage` component.
+- [Marek Kurańda](https://github.com/mjkuranda): Creator for `offers` at shop subpage.
 
 ## [0.20.0] - 2024-05-14
 ### Added