+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `Loader` component.
 - [Marek Kurańda](https://github.com/mjkuranda): `Loader`s to appropriate places in app.
+- [Marek Kurańda](https://github.com/mjkuranda): `SettingsCreatorContext` to store loading state while submitting any creator form.
 
 ## [0.21.0] - 2024-05-15
 ### Added